+++ configv2/example_factories.go
@@ -17,13 +17,15 @@ package configv2
 import (
 	"context"
 
+	"go.uber.org/zap"
+
 	"github.com/open-telemetry/opentelemetry-service/consumer"
 	"github.com/open-telemetry/opentelemetry-service/data"
+	"github.com/open-telemetry/opentelemetry-service/exporter"
 	"github.com/open-telemetry/opentelemetry-service/factories"
 	"github.com/open-telemetry/opentelemetry-service/models"
 	"github.com/open-telemetry/opentelemetry-service/processor"
 	"github.com/open-telemetry/opentelemetry-service/receiver"
-	"go.uber.org/zap"
 )
 
 // ExampleReceiver is for testing purposes. We are defining an example config and factory
@@ -252,12 +254,12 @@ func (f *ExampleExporterFactory) CreateDefaultConfig() models.Exporter {
 }
 
 // CreateTraceExporter creates a trace exporter based on this config.
-func (f *ExampleExporterFactory) CreateTraceExporter(logger *zap.Logger, cfg models.Exporter) (consumer.TraceConsumer, factories.StopFunc, error) {
+func (f *ExampleExporterFactory) CreateTraceExporter(logger *zap.Logger, cfg models.Exporter) (consumer.TraceConsumer, exporter.StopFunc, error) {
 	return &ExampleExporterConsumer{}, nil, nil
 }
 
 // CreateMetricsExporter creates a metrics exporter based on this config.
-func (f *ExampleExporterFactory) CreateMetricsExporter(logger *zap.Logger, cfg models.Exporter) (consumer.MetricsConsumer, factories.StopFunc, error) {
+func (f *ExampleExporterFactory) CreateMetricsExporter(logger *zap.Logger, cfg models.Exporter) (consumer.MetricsConsumer, exporter.StopFunc, error) {
 	return &ExampleExporterConsumer{}, nil, nil
 }
 
@@ -327,7 +329,7 @@ func (f *ExampleProcessorFactory) CreateMetricsProcessor(
 func RegisterTestFactories() error {
 	_ = receiver.RegisterReceiverFactory(&ExampleReceiverFactory{})
 	_ = receiver.RegisterReceiverFactory(&MultiProtoReceiverFactory{})
-	_ = factories.RegisterExporterFactory(&ExampleExporterFactory{})
+	_ = exporter.RegisterExporterFactory(&ExampleExporterFactory{})
 	_ = factories.RegisterProcessorFactory(&ExampleProcessorFactory{})
 	return nil
 }
