+++ exporter/prometheusexporter/prometheus_test.go
@@ -28,8 +28,8 @@ import (
 	metricspb "github.com/census-instrumentation/opencensus-proto/gen-go/metrics/v1"
 
 	"github.com/open-telemetry/opentelemetry-service/data"
-	"github.com/open-telemetry/opentelemetry-service/factories"
-	viperutils "github.com/open-telemetry/opentelemetry-service/internal/config/viperutils"
+	"github.com/open-telemetry/opentelemetry-service/exporter"
+	"github.com/open-telemetry/opentelemetry-service/internal/config/viperutils"
 )
 
 func TestPrometheusExporter(t *testing.T) {
@@ -53,7 +53,7 @@ func TestPrometheusExporter(t *testing.T) {
 		},
 	}
 
-	factory := factories.GetExporterFactory(typeStr)
+	factory := exporter.GetExporterFactory(typeStr)
 	for i, tt := range tests {
 		// Run it a few times to ensure that shutdowns exit cleanly.
 		for j := 0; j < 3; j++ {
@@ -105,7 +105,7 @@ func TestPrometheusExporter_endToEnd(t *testing.T) {
 		Endpoint: ":7777",
 	}
 
-	factory := factories.GetExporterFactory(typeStr)
+	factory := exporter.GetExporterFactory(typeStr)
 	consumer, stopFunc, err := factory.CreateMetricsExporter(zap.NewNop(), config)
 	assert.Nil(t, err)
